Разбил код на модули, добавил строку для работы с модулями в package.json
