+++ package.json
@@ -23,5 +23,7 @@
   },
   "engines": {
     "node": ">=12.0.0"
-  }
+  },
+
+  "type": "module"
 }
